requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
sqlparse>=0.4.0
chromadb>=0.4.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
anthropic>=0.39.0